requirements.txt
Flask==2.1.1
requests==2.26.0
python-dotenv==0.19.2
authlib==0.15.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.31.0
google-api-core==2.3.2
